Replace libopus with new Opus encoder

+++ CM4/Core/Inc/opus_config.h
@@ -7,7 +7,7 @@ struct __attribute__ ((__packed__, aligned (4))) opus_frame_t;
 
 typedef struct __attribute__ ((__packed__, aligned (4))) opus_frame_t
 {
-   uint16_t frame_delimiter, num_encoded_bytes;
+   uint8_t frame_delimiter, num_encoded_bytes;
    uint8_t encoded_data[2 * ((OPUS_ENCODED_BIT_RATE/8) / (1000/OPUS_MS_PER_FRAME))];
    struct opus_frame_t *next;
 } opus_frame_t;
